Adding fully operational setup.py file.
